loudml/server: implement pool/worker/job consuming mechanism
Long operations such as training and prediction must be performed by
subprocesses.

Each subprocess run a Worker and its code is located in worker.py.
Subprocceses are managed via a Pool object.

When a long operation is requested, a Job object is instanciated and
submitted to the Pool.

When a Worker starts consuming a job, its sends a message to the main
process via a Queue object and returns the result when it's done.

In the future the Queue object could also be used for updating jobs
progression (10%, 20%, 30%, ...).

Job statuses are available as a resource through the /jobs endpoint.

Signed-off-by: Vianney Bajart <vianney@redmintnetwork.fr>