timeseries: allow to use previous value as default value
When no data are found for a given bucket, it's sometime useful to fill it
with the last known value instead of None or 0.

To configure this, just set 'previous' in the 'default' field of the
feature.

Example:

{
    "name": "foo",
    "metric": "avg",
    "field": "foo",
    "default": "previous",
}

Note: if the first bucket is empty, the value will be None anyway.
Signed-off-by: Vianney Bajart <vianney@redmintnetwork.fr>