warp10: add a 'global_prefix' settings + use collection-like prefix
This optional 'global_prefix' is configured on the datasource and can be
used to add a prefix to all the metrics names/selectors in the queries
send to the Warp10 server.

Warp10 has no collection/measurement/table concept. Instead, the user can
simply add a prefix to feature fields.

Example:

mycollection.field1
mycollection.field2

Signed-off-by: Vianney Bajart <vianney@redmintnetwork.fr>