fix date format that depends on the index mapping if not forced in the query
Internally, a date is represented as a 64 bit number representing a timestamp
in milliseconds-since-the-epoch. These timestamps are returned as the bucket keys.
The key_as_string is the same timestamp converted to a formatted date string using
the format specified with the format parameter:

If no format is specified, then it will use the first date format specified in
the field mapping.